pyproject.toml
[project]
name = "fastapi-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.16.2",
    "asyncpg>=0.30.0",
    "fastapi>=0.115.13",
    "pre-commit>=4.2.0",
    "pydantic-settings>=2.10.0",
    "ruff>=0.12.0",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.3",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]
fixable = ["ALL"]


[tool.ruff.format]
quote-style = "double"          # или 'single'

